Added more entity states and also modify how the display buffer is rendered on.I also added ground collision system and component.

+++ src/main/java/bolt/World.java
@@ -79,7 +79,7 @@ public class World{
   }
 
 
-  public void update(long dt){
+  public void update(float dt){
     systemManager.update(dt);
   }
